results/publication_ready_package_20260430_073301/requirements.txt
transformers>=4.40.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0.0
sentencepiece>=0.1.99
